+++ src/configType.ts
