+++ src/database/historicalPairs.ts
