+++ src/utils/dayOfWeekTranslation.ts
